Move requiredFields to vuex
